lib: split stone library standalone from parsing library
