+++ src/shiratsu-lib/Cargo.toml
@@ -21,14 +21,16 @@ phf = { version = "0.8", features = ["macros"] }
 # Serde
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-quick-xml = { version = "0.18", features = [ "serialize" ] }
+quick-xml = { version = "0.22", features = [ "serialize" ] }
 
 # Parsing
-nom = { version = "6", features = ["regexp"] }
+nom = { version = "6" }
 indexmap = "1.3"
-regex = "1"
-fancy-regex = "0.3" # tosec lies.
 array_iterator = "0.2"
+
+# todo: only enable when using legacy parsers
+regex = "1"
+fancy-regex = "0.5" # tosec lies.
 
 # pin funty
 funty = "=1.1.0"
